Add evaluation of CAdES Baseline-T timeStampToken
The timestamp token is embedded as an unsigned attribute.
It states a time on which the signature was available and can therefore
be used as a reference time that could provide more trust thant the
time being just part of the signature.
If it is avialble and -cades mode is set, it is then used for the
verification.