Fix RECORD_LAYER_write_pending wrong result with pipelining
when using write pipelines it is possible that there is not
enough data to fill all write buffers. In that case it is
pssible that empty write buffers are left at the end. However
RECORD_LAYER_write_pending did not handle that case properly.
That may cause data loss after a partial write, due to the
wrong result, because ssl3_write_pending will not be called,
and thus the pending data will be lost.

Fixes: #21461