fix leaked pointer when BIO_NOCLOSE is set
Was working on an application when valgrind generated this memory leak
error for us:

==1007580== at 0x483C815: malloc (vg_replace_malloc.c:431)
==1007580== by 0x2C2689: CRYPTO_zalloc (in /home/vien/microedge-c/test)
==1007580== by 0x295A17: BUF_MEM_new (in /home/vien/microedge-c/test)
==1007580== by 0x295A78: BUF_MEM_new_ex (in /home/vien/microedge-c/test)
==1007580== by 0x28CACE: mem_new (in /home/vien/microedge-c/test)
==1007580== by 0x285EA8: BIO_new_ex (in /home/vien/microedge-c/test)
==1007580== by 0x231894: convert_pubkey_ECC (tpm2_driver.c:221)
==1007580== by 0x232B73: create_ephemeral_key (tpm2_driver.c:641)
==1007580== by 0x232E1F: tpm_gen_keypair (tpm2_driver.c:695)
==1007580== by 0x22D60A: gen_keypair (se_driver_api.c:275)
==1007580== by 0x21FF35: generate_keypair (dhkey.c:142)
==1007580== by 0x24D4C8: __test_dhkey (dhkey_test.c:55)

The error occurs because BIO_new(BIO_s_mem()) allocates:
1) a BIO struct
2) a BIO_BUF_MEM struct which holds a data pointer that is
   (re)-allocated as data is written to the BIO

When BIO_set_close(bio, BIO_NOCLOSE) is set in the calling program, the
expectation is that the BIO_BUF_MEM data pointer is orhpaned so that the
calling code can take ownership of it via BIO_get_mem_data().

This all works as expected.  However, on calling BIO_free, the s_mem
implementation calls mem_free->mem_buf_free which, when BIO_NOCLOSE is
set (shadowed via BIO->shutdown), implements the data pointer orphaning
by just not calling BIO_BUF_free.  This properly skips the freeing of
the data pointer, but also leaves the controlling BIO_BUF_MEM struct
allocated, leading to the leak report above.

Unfortunately, this is complicated by the semantics of the API.
BIO_set_close flags the BIO to free the BUF_MEM structure (or not,
depending on weather BIO_CLOSE or BIO_NOCLOSE is set, but makes no
determination as to the status of the internal data pointer, meaning
depending on the application behavior, the value returned by
BIO_get_mem_data may or may not be freed, leading to leaked memory or
double frees.

So I'm proposing this two point solution

1) A call to BIO_get_mem_data implies an orphaning of the BUF_MEM data
   pointer.  That is to say, calling BIO_get_mem_data tells openssl that
   the caller has a pointer to the data, and now owns it.  On free,
   regardless of the BIO_CLOSE/BIO_NOCLOSE state (bio->shutdown), the
   data pointer is orphaned, and it becomes the responsibility of the
   caller to free it.  This may result in memory leaks for callers who
   call BIO_get_mem_data, BIO_get_mem_ptr, BIO_set_close(bio, BIO_NOCLOSE)
   and BUF_MEM_free, but here we are trading one leak for another, and
   its a leak that applications can manage, rather than an internal leak
   that they have no purview over.

2) Raise an error from BIO_free, if the BIO_s_mem buf underlying hasn't
   noted that BIO_get_mem_ptr has been called with BIO_set_close(bio,
   BIO_NOCLOSE) set.  In this case, the library has noted that it is
   being asked to orphan a BUF_MEM structure, but the applicatoin is
   guaranteed not to have a pointer to said BUF_MEM to free independently.
   Properly written applications should check the return of BIO_free,
   see the implied fetch error, and take appropriate corrective
   action (namely, get the BUF_MEM ptr, or clear NOCLOSE, and try the
   free again).

Implementing this proposal requires the addition of two new flags in the
bio BIO_FLAGS_MEMBUF_FETCH and BIO_FLAGS_DATA_FETCH, which are managed
internally.  I'm not a huge fan of adding those, but we appear to have
space in the flags field, and it means we can remain ABI compatible,
which is good, as it allows for a better API changing fix in the future