Remove X9.31 support from FIPS provider
This patch is a part of Red Hat FIPS-related patches

The current draft of FIPS 186-5 [1] no longer contains specifications
for X9.31 signature padding. Instead, it contains the following
information in Appendix E:

> ANSI X9.31 was withdrawn, so X9.31 RSA signatures were removed from
> this standard.

Since this situation is unlikely to change in future revisions of the
draft, and future FIPS 140-3 validations of the provider will require
X9.31 to be disabled or marked as not approved with an explicit
indicator, disallow this padding mode now.

Remove the X9.31 tests from the acvp test, since they will always fail
now.

 [1]: https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.186-5-draft.pdf

the code is available here
https://gitlab.com/redhat/centos-stream/rpms/openssl/-/blob/c9s/0081-signature-Remove-X9.31-padding-from-FIPS-prov.patch?ref_type=heads