Use combination of DigestSign/DigestVerify functions in FIPS KATs
This patch is a part of Red Hat FIPS-related patches

In review for FIPS 140-3, the lack of a self-test for the digest_sign
and digest_verify provider functions was highlighted as a problem. NIST
no longer provides ACVP tests for the RSA SigVer primitive (see
https://github.com/usnistgov/ACVP/issues/1347). Because FIPS 140-3
recommends the use of functions that compute the digest and signature
within the module, we have been advised in our module review that the
self tests should also use the combined digest and signature APIs, i.e.
the digest_sign and digest_verify provider functions.

Modify the signature self-test to use these instead by switching to
EVP_DigestSign and EVP_DigestVerify. This requires adding more ifdefs to
crypto/evp/m_sigver.c to make these functions usable in the FIPS module.

the code is available here
https://gitlab.com/redhat/centos-stream/rpms/openssl/-/blob/c9s/0074-FIPS-Use-digest_sign-digest_verify-in-self-test.patch?ref_type=heads