amend! Fix a failure in bio_dgram_test on the NonStop platform
Fix a failure in bio_dgram_test on the NonStop platform

The size of the datagram header is significantly larger that we might
expect on NonStop (probably driven by sizeof(BIO_ADDR)). We adjust the
size of the default buffer to take into account the header size and the
mtu.

Fixes #22013