Clean up the blake2 provider implementation
XOFLEN was only implemented for BLAKE2b, not for BLAKE2s.

Also, it was made in such a way that a call of the set_ctx_params function
to set XOFLEN only set the BLAKE2[SB]_PARAM structure, but never had an
effect on the outlen field of the corresponding BLAKE2[SB]_CTX structure,
which effectively ignored whatever was set.

Finally, it's cleaner with a provider implementation that uses
IMPLEMENT_digest_functions_with_settable_ctx(), rather than reimplementing
everything that macro does.