Correct documentation for PKCS5_PBKDF2_HMAC in OpenSSL 3.0
In OpenSSL 3.0, the documentation for PKCS5_PBKDF2_HMAC incorrectly states
that an iter value less than 1 is treated as a single iteration. Upon further
investigation in providers/implementations/kdfs/pbkdf2.c, it appears that
invalid iter values will result in failure and raise the
PROV_R_INVALID_ITERATION_COUNT error. This commit corrects the documentation
to accurately reflect the behavior in OpenSSL 3.0.

Closes openssl#22168

Signed-off-by: Sumitra Sharma <sumitraartsy@gmail.com>