windows CE: clarify conditional compilation (should be no-op)
The condition
  #if !defined(OPENSSL_SYS_WINCE)
was changed to
  #if !defined(OPENSSL_SYS_WINCE) && !defined(_WIN32_WCE)

Which is, hopefully, a no-op, but the change makes it clearer
that this block of code is not entered if _WIN32_WCE is defined.

If it could enter before this commit, then it would have failed to
compile on CE due to the new variable h_conin which is declared
behind #if !defined(_WIN32_WCE), but used in this block of code.

This conditional block of code uses ReadConsoleW to read UTF-8 input
if supported, and ReadConsoleA otherwise.

I don't have a CE build system to test this, but I think _WIN32_WCE
is never defined without OPENSSL_SYS_WINCE (hence no-op), because:

- It feels logical that OPENSSL_SYS_WINCE and _WIN32_WCE are defined
  together and generally have the same effect in this context.

- At Configurations/10-main.conf, -D_WIN32_WCE=... is set only at
  vc_wince_info, which is used only at the "VC-CE" block, which also
  sets `sys_id => "WINCE"`, which likely becomes OPENSSL_SYS_WINCE,
  hence they're probably always defined together.

- The commit which introduced the usage of console APIs - 087d3e89
  puts all the code behind #if !defined(_WIN32_WCE), except inside
  "our" block - which is already in #if !defined(OPENSSL_SYS_WINCE),
  suggesting that this code is not intended for Windows CE.
  Also, _WIN32_WCE was not used at this file before this commit.
  Also, it's unknown whether it was tested with CE when it was added.

- There are 17 source files which use _WIN32_WCE, and 8 which use
  OPENSSL_SYS_WINCE, but excluding our file (ui_openssl.c), only two
  other files have both (sockets.h, e_os.h), suggesting that typically
  they're used interchangably, as an arbitrary choice by the author,
  but headers can have many authors, hence possibly inconsistent use.

- OPENSSL_SYS_WINCE exists at the code since commit 0bf23d9b, which
  introced CE support. _WIN32_WCE first appears at the code 3 years
  later, and in one of its first appearances it's used to test
  specific CE versions - commit 8a35fb3b , suggesting it's only
  a specific case of OPENSSL_SYS_WINCE (hence this commit is no-op).

- Searching for CE docs (hard!), I couldn't find ReadConsole{A,W}
  which are used in this block, which supports the case that it
  already never intended to be used with _WIN32_WCE defined.

Conclusion:
While not definitive, to me the above suggest that _WIN32_WINCE
is only defined when OPENSSL_SYS_WINCE is defined, hence this commit
is indeed a no-op, but still useful for clarity.

If this block did, in fact, enter with _WIN32_WCE defined, then it
means this commit is not no-op, and would prevent using the console
APIs to read user input on CE, and possibly also prevent Unicode
input (if CP_UTF8 is defined on CE - which I also don't know).

However, it won't break no-echo for password input, because the code
which disables echo was already excluded when _WIN32_WCE is defined.

To restore the removed functionality, if this commit is not no-op,
we'd need to ensure that h_conin is defined for _WIN32_WCE, that
it's initialized correctly (probably using STD_INPUT_HANDLE), and
that is_a_tty is kept in sync where appropriate.

Note:
This is way too much work only to identify whether code is entered or
not entered in some platform, and ultimately it still ended up as an
assessment and not a definitive conclusion, and who knows if anyone
would actually even try to build for Windows CE in the future...

CLA: trivial