ecc: Common bi-endian fallback routines for nistp curves
The Solinas' implementations for nistp curves, enabled with the
ec-nistp_128_gcc_64 compile flag, are currently gated behind the
requirement of a little-endian target. This is due to the fact that each
of these implementations will generate little-endian byte arrays of
precomputed values for use in a wNAF ladder.

Conversion to and from this common little-endian byte array format does
not occur in a fast path, and so provide two simple routines for
converting internal representations to and from this format.

 - ec_nistp_pre_comp_deserialize()
 - ec_nistp_pre_comp_serialize()

These routines are intended just for use as a fallback for big-endian
systems, but should generate correct code on any endianness.

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>