ciphers: Reject update if module isn't operational
When a FIPS self-test such as for example a pair-wise consistency test
fails, the module should go into an error state and refuse any further
cryptographic operations.

This is not the case for the RSA or ECDSA pair-wise consistency test,
after which EVP_EncryptUpdate() can still be called.

Add invocations of ossl_prov_is_running() to provider functions called
by EVP_EncryptUpdate() to fix this.

Signed-off-by: Clemens Lang <cllang@redhat.com>